+++ examples/confirm.rs
@@ -1,30 +1,70 @@
-extern crate dialoguer;
-
-use dialoguer::Confirm;
+use dialoguer::{theme::ColorfulTheme, Confirm};
 
 fn main() {
-    println!("disable default");
-    if Confirm::new()
-        .with_prompt("continue?")
-        .disable_default(true)
+    if Confirm::with_theme(&ColorfulTheme::default())
+        .with_prompt("Do you want to continue?")
         .interact()
         .unwrap()
     {
-        println!("continuing");
+        println!("Looks like you want to continue");
     } else {
-        println!("exiting");
+        println!("nevermind then :(");
     }
 
-    println!();
-
-    println!("enable default");
-    if Confirm::new()
-        .with_prompt("continue?")
+    if Confirm::with_theme(&ColorfulTheme::default())
+        .with_prompt("Do you really want to continue?")
+        .default(true)
         .interact()
         .unwrap()
     {
-        println!("continuing");
+        println!("Looks like you want to continue");
     } else {
-        println!("exiting");
+        println!("nevermind then :(");
+    }
+
+    if Confirm::with_theme(&ColorfulTheme::default())
+        .with_prompt("Do you really really want to continue?")
+        .default(true)
+        .show_default(false)
+        .wait_for_newline(true)
+        .interact()
+        .unwrap()
+    {
+        println!("Looks like you want to continue");
+    } else {
+        println!("nevermind then :(");
+    }
+
+    if Confirm::with_theme(&ColorfulTheme::default())
+        .with_prompt("Do you really really really want to continue?")
+        .wait_for_newline(true)
+        .interact()
+        .unwrap()
+    {
+        println!("Looks like you want to continue");
+    } else {
+        println!("nevermind then :(");
+    }
+
+    match Confirm::with_theme(&ColorfulTheme::default())
+        .with_prompt("Do you really really really really want to continue?")
+        .interact_opt()
+        .unwrap()
+    {
+        Some(true) => println!("Looks like you want to continue"),
+        Some(false) => println!("nevermind then :("),
+        None => println!("Ok, we can start over later"),
+    }
+
+    match Confirm::with_theme(&ColorfulTheme::default())
+        .with_prompt("Do you really really really really really want to continue?")
+        .default(true)
+        .wait_for_newline(true)
+        .interact_opt()
+        .unwrap()
+    {
+        Some(true) => println!("Looks like you want to continue"),
+        Some(false) => println!("nevermind then :("),
+        None => println!("Ok, we can start over later"),
     }
 }