+++ src/prompts/confirm.rs
@@ -2,14 +2,14 @@ use std::io;
 
 use crate::theme::{SimpleTheme, TermThemeRenderer, Theme};
 
-use console::Term;
+use console::{Key, Term};
 
 /// Renders a confirm prompt.
 ///
 /// ## Example usage
 ///
 /// ```rust,no_run
-/// # fn test() -> Result<(), Box<std::error::Error>> {
+/// # fn test() -> Result<(), Box<dyn std::error::Error>> {
 /// use dialoguer::Confirm;
 ///
 /// if Confirm::new().with_prompt("Do you want to continue?").interact()? {
@@ -21,9 +21,8 @@ use console::Term;
 /// ```
 pub struct Confirm<'a> {
     prompt: String,
-    default: bool,
+    default: Option<bool>,
     show_default: bool,
-    disable_default: bool,
     wait_for_newline: bool,
     theme: &'a dyn Theme,
 }
@@ -41,14 +40,14 @@ impl<'a> Confirm<'a> {
     }
 
     /// Creates a confirm prompt with a specific theme.
-    /// 
+    ///
     /// ## Examples
     /// ```rust,no_run
     /// use dialoguer::{
     ///     Confirm,
     ///     theme::ColorfulTheme
     /// };
-    /// 
+    ///
     /// # fn main() -> std::io::Result<()> {
     /// let proceed = Confirm::with_theme(&ColorfulTheme::default())
     ///     .with_prompt("Do you wish to continue?")
@@ -59,9 +58,8 @@ impl<'a> Confirm<'a> {
     pub fn with_theme(theme: &'a dyn Theme) -> Confirm<'a> {
         Confirm {
             prompt: "".into(),
-            default: true,
+            default: None,
             show_default: true,
-            disable_default: false,
             wait_for_newline: false,
             theme,
         }
@@ -92,32 +90,22 @@ impl<'a> Confirm<'a> {
         self.wait_for_newline = wait;
         self
     }
-    /// Overrides the default output if user pushes enter key without inputing any character.
-    /// Character corresponding to the default choice (e.g `Y` if default is `true`) will be uppercased in the displayed prompt.
-    /// 
-    /// The default output is true.
+
+    /// Sets a default.
+    ///
+    /// Out of the box the prompt does not have a default and will continue
+    /// to display until the user inputs something and hits enter. If a default is set the user
+    /// can instead accept the default with enter.
     pub fn default(&mut self, val: bool) -> &mut Confirm<'a> {
-        self.default = val;
+        self.default = Some(val);
         self
     }
 
-    /// Disables or enables display of options user can choose from.
+    /// Disables or enables the default value display.
     ///
-    /// The default is to append `[y/n]` to the prompt to tell the
-    /// user which keys to press. This also renders the default choice
-    /// in uppercase. The default is selected on enter.
+    /// The default is to append the default value to the prompt to tell the user.
     pub fn show_default(&mut self, val: bool) -> &mut Confirm<'a> {
         self.show_default = val;
-        self
-    }
-
-    /// Select whether the user must explicitly confirm their selection.
-    ///
-    /// When `false` (default), the user can input a newline to select the default.
-    ///
-    /// When `true`, the user must input a letter - a newline will do nothing.
-    pub fn disable_default(&mut self, val: bool) -> &mut Confirm<'a> {
-        self.disable_default = val;
         self
     }
 
@@ -125,20 +113,29 @@ impl<'a> Confirm<'a> {
     ///
     /// If the user confirms the result is `true`, `false` if declines or default (configured in [default](#method.default)) if pushes enter.
     /// Otherwise function discards input waiting for valid one.
-    /// 
+    ///
     /// The dialog is rendered on stderr.
     pub fn interact(&self) -> io::Result<bool> {
         self.interact_on(&Term::stderr())
     }
 
+    /// Enables user interaction and returns the result.
+    ///
+    /// This method is similar to [interact_on_opt](#method.interact_on_opt) except for the fact that it does not allow selection of the terminal.
+    /// The dialog is rendered on stderr.
+    /// Result contains `Some(bool)` if user answered "yes" or "no" or `None` if user cancelled with 'Esc' or 'q'.
+    pub fn interact_opt(&self) -> io::Result<Option<bool>> {
+        self.interact_on_opt(&Term::stderr())
+    }
+
     /// Like [interact](#method.interact) but allows a specific terminal to be set.
-    /// 
-    /// ## Examples 
-    /// 
-    /// ```rust,nor_run
+    ///
+    /// ## Examples
+    ///
+    /// ```rust,no_run
     /// use dialoguer::Confirm;
     /// use console::Term;
-    /// 
+    ///
     /// # fn main() -> std::io::Result<()> {
     /// let proceed = Confirm::new()
     ///     .with_prompt("Do you wish to continue?")
@@ -147,62 +144,111 @@ impl<'a> Confirm<'a> {
     /// # }
     /// ```
     pub fn interact_on(&self, term: &Term) -> io::Result<bool> {
+        self._interact_on(term, false)?
+            .ok_or_else(|| io::Error::new(io::ErrorKind::Other, "Quit not allowed in this case"))
+    }
+
+    /// Like [interact_opt](#method.interact_opt) but allows a specific terminal to be set.
+    ///
+    /// ## Examples
+    /// ```rust,no_run
+    /// use dialoguer::Confirm;
+    /// use console::Term;
+    ///
+    /// fn main() -> std::io::Result<()> {
+    ///     let confirmation = Confirm::new()
+    ///         .interact_on_opt(&Term::stdout())?;
+    ///
+    ///     match confirmation {
+    ///         Some(answer) => println!("User answered {}", if answer { "yes" } else { "no " }),
+    ///         None => println!("User did not answer")
+    ///     }
+    ///
+    ///     Ok(())
+    /// }
+    /// ```
+    pub fn interact_on_opt(&self, term: &Term) -> io::Result<Option<bool>> {
+        self._interact_on(term, true)
+    }
+
+    fn _interact_on(&self, term: &Term, allow_quit: bool) -> io::Result<Option<bool>> {
         let mut render = TermThemeRenderer::new(term, self.theme);
 
-        let default = if self.show_default {
-            Some(self.default)
+        let default_if_show = if self.show_default {
+            self.default
         } else {
             None
         };
-        render.confirm_prompt(&self.prompt, default)?;
+
+        render.confirm_prompt(&self.prompt, default_if_show)?;
 
         term.hide_cursor()?;
         term.flush()?;
 
+        let rv;
+
         if self.wait_for_newline {
             // Waits for user input and for the user to hit the Enter key
             // before validation.
-            let mut input_buf = String::new();
+            let mut value = default_if_show;
+
             loop {
-                io::stdin().read_line(&mut input_buf)?;
-                let rv = match &*input_buf.trim_end().to_lowercase() {
-                    "y" | "yes" => true,
-                    "n" | "no" => false,
-                    "" if !self.disable_default => self.default,
-                    _ => {
-                        // On invalid input re-render the user prompt.
-                        render.confirm_prompt(&self.prompt, default)?;
-                        input_buf.clear();
+                let input = term.read_key()?;
+
+                match input {
+                    Key::Char('y') | Key::Char('Y') => {
+                        value = Some(true);
+                    }
+                    Key::Char('n') | Key::Char('N') => {
+                        value = Some(false);
+                    }
+                    Key::Enter => {
+                        if !allow_quit {
+                            value = value.or(self.default);
+                        }
+
+                        if value.is_some() || allow_quit {
+                            rv = value;
+                            break;
+                        }
                         continue;
                     }
-                };
-
-                term.show_cursor()?;
-                term.flush()?;
-
-                return Ok(rv);
-            }
-        } else {
-            // Default behavior: matches continuously on every keystroke,
-            // and does not wait for user to hit the Enter key.
-            loop {
-                let input = term.read_char()?;
-                let rv = match input {
-                    'y' | 'Y' => true,
-                    'n' | 'N' => false,
-                    '\n' | '\r' if !self.disable_default => self.default,
+                    Key::Escape | Key::Char('q') if allow_quit => {
+                        value = None;
+                    }
                     _ => {
                         continue;
                     }
                 };
 
                 term.clear_line()?;
-                render.confirm_prompt_selection(&self.prompt, rv)?;
-                term.show_cursor()?;
-                term.flush()?;
+                render.confirm_prompt(&self.prompt, value)?;
+            }
+        } else {
+            // Default behavior: matches continuously on every keystroke,
+            // and does not wait for user to hit the Enter key.
+            loop {
+                let input = term.read_key()?;
+                let value = match input {
+                    Key::Char('y') | Key::Char('Y') => Some(true),
+                    Key::Char('n') | Key::Char('N') => Some(false),
+                    Key::Enter if self.default.is_some() => Some(self.default.unwrap()),
+                    Key::Escape | Key::Char('q') if allow_quit => None,
+                    _ => {
+                        continue;
+                    }
+                };
 
-                return Ok(rv);
+                rv = value;
+                break;
             }
         }
+
+        term.clear_line()?;
+        render.confirm_prompt_selection(&self.prompt, rv)?;
+        term.show_cursor()?;
+        term.flush()?;
+
+        Ok(rv)
     }
 }