+++ src/theme.rs
@@ -28,7 +28,7 @@ pub trait Theme {
             write!(f, "{} ", &prompt)?;
         }
         match default {
-            None => {}
+            None => write!(f, "[y/n] ")?,
             Some(true) => write!(f, "[Y/n] ")?,
             Some(false) => write!(f, "[y/N] ")?,
         }
@@ -40,12 +40,21 @@ pub trait Theme {
         &self,
         f: &mut dyn fmt::Write,
         prompt: &str,
-        selection: bool,
+        selection: Option<bool>,
     ) -> fmt::Result {
-        if prompt.is_empty() {
-            write!(f, "{}", if selection { "yes" } else { "no" })
-        } else {
-            write!(f, "{} {}", &prompt, if selection { "yes" } else { "no" })
+        let selection = selection.map(|b| if b { "yes" } else { "no" });
+
+        match selection {
+            Some(selection) if prompt.is_empty() => {
+                write!(f, "{}", selection)
+            }
+            Some(selection) => {
+                write!(f, "{} {}", &prompt, selection)
+            }
+            None if prompt.is_empty() => Ok(()),
+            None => {
+                write!(f, "{}", &prompt)
+            }
         }
     }
 
@@ -352,18 +361,23 @@ impl Theme for ColorfulTheme {
         }
 
         match default {
-            None => write!(f, "{}", &self.prompt_suffix),
+            None => write!(
+                f,
+                "{} {}",
+                self.hint_style.apply_to("(y/n)"),
+                &self.prompt_suffix
+            ),
             Some(true) => write!(
                 f,
                 "{} {} {}",
-                self.hint_style.apply_to("(Y/n)"),
+                self.hint_style.apply_to("(y/n)"),
                 &self.prompt_suffix,
                 self.defaults_style.apply_to("yes")
             ),
             Some(false) => write!(
                 f,
                 "{} {} {}",
-                self.hint_style.apply_to("(y/N)"),
+                self.hint_style.apply_to("(y/n)"),
                 &self.prompt_suffix,
                 self.defaults_style.apply_to("no")
             ),
@@ -375,7 +389,7 @@ impl Theme for ColorfulTheme {
         &self,
         f: &mut dyn fmt::Write,
         prompt: &str,
-        selection: bool,
+        selection: Option<bool>,
     ) -> fmt::Result {
         if !prompt.is_empty() {
             write!(
@@ -385,14 +399,21 @@ impl Theme for ColorfulTheme {
                 self.prompt_style.apply_to(prompt)
             )?;
         }
+        let selection = selection.map(|b| if b { "yes" } else { "no" });
 
-        write!(
-            f,
-            "{} {}",
-            &self.success_suffix,
-            self.values_style
-                .apply_to(if selection { "yes" } else { "no" })
-        )
+        match selection {
+            Some(selection) => {
+                write!(
+                    f,
+                    "{} {}",
+                    &self.success_suffix,
+                    self.values_style.apply_to(selection)
+                )
+            }
+            None => {
+                write!(f, "{}", &self.success_suffix)
+            }
+        }
     }
 
     /// Formats an input prompt after selection.
@@ -642,7 +663,7 @@ impl<'a> TermThemeRenderer<'a> {
         self.write_formatted_str(|this, buf| this.theme.format_confirm_prompt(buf, prompt, default))
     }
 
-    pub fn confirm_prompt_selection(&mut self, prompt: &str, sel: bool) -> io::Result<()> {
+    pub fn confirm_prompt_selection(&mut self, prompt: &str, sel: Option<bool>) -> io::Result<()> {
         self.write_formatted_prompt(|this, buf| {
             this.theme.format_confirm_prompt_selection(buf, prompt, sel)
         })